FPVTL-497: 'Review RA request' task update
