+++ src/functionalTest/java/uk/gov/hmcts/reform/prl/controllers/caseflags/CaseFlagsControllerFunctionalTest.java
@@ -40,7 +40,20 @@ public class CaseFlagsControllerFunctionalTest {
     private final RequestSpecification request = RestAssured.given().relaxedHTTPSValidation().baseUri(targetInstance);
 
     @Test
-    public void givenRequestBody_whenAboutToStart_then200Response() throws Exception {
+    public void givenBodyWithCaseFlags_whenAboutToStartCallback() throws Exception {
+        String requestBody = ResourceLoader.loadJson(VALID_REQUEST_BODY_WITH_FLAGS);
+        request
+            .header("Authorization", idamTokenGenerator.generateIdamTokenForSystem())
+            .body(requestBody)
+            .when()
+            .contentType("application/json")
+            .post("/caseflags/about-to-start")
+            .then()
+            .assertThat().statusCode(200);
+    }
+
+    @Test
+    public void givenRequestBodyReviewAndSm_whenAboutToStart_then200Response() throws Exception {
         String requestBody = ResourceLoader.loadJson(VALID_REQUEST_BODY_INITIAL);
         String client = ResourceLoader.loadJson(CLIENT_CONTEXT_CASE_NOTE);
         request
@@ -56,7 +69,21 @@ public class CaseFlagsControllerFunctionalTest {
     }
 
     @Test
-    public void givenRequestBody_whenAboutToSubmit_then200Response() throws Exception {
+    public void givenBodyWithCaseFlags_whenSubmitted() throws Exception {
+        String requestBody = ResourceLoader.loadJson(VALID_REQUEST_BODY_WITH_FLAGS);
+        request
+            .header("Authorization", idamTokenGenerator.generateIdamTokenForSystem())
+            .body(requestBody)
+            .when()
+            .contentType("application/json")
+            .post("/caseflags/submitted")
+            .then()
+            .assertThat().statusCode(200);
+
+    }
+
+    @Test
+    public void givenRequestBodyReviewAndSm_whenAboutToSubmit_then200Response() throws Exception {
         String requestBody = ResourceLoader.loadJson(VALID_REQUEST_BODY_WITH_FLAGS);
         request
             .header("Authorization", idamTokenGenerator.generateIdamTokenForCourtAdmin())
@@ -70,7 +97,7 @@ public class CaseFlagsControllerFunctionalTest {
     }
 
     @Test
-    public void givenRequestBody_whenSetUpWaTest_then200Response() throws Exception {
+    public void givenRequestBodyReviewAndSm_whenSetUpWaTest_then200Response() throws Exception {
         String requestBody = ResourceLoader.loadJson(VALID_REQUEST_BODY_WITH_FLAGS);
         request
             .header("Authorization", idamTokenGenerator.generateIdamTokenForCourtAdmin())
