+++ src/functionalTest/resources/requests/caseFlags/review-lang-sm-initial.json
@@ -45,7 +45,9 @@
             "dateCreated": "2025-06-02T16:27:44.593764142"
           }
         }
-      ]
+      ],
+      "isCaseFlagsTaskCreated": "No",
+      "selectedFlags": []
     }
   }
 }
